Домашнее задание     Автоматическое логирование
