+++ hw06-gradle/src/main/java/ru/otus/hw06/interfaces/ATMVisitor.java
@@ -1,0 +1,8 @@
+package ru.otus.hw06.interfaces;
+
+import ru.otus.hw06.exceptions.ATMExceptions;
+
+public interface ATMVisitor {
+
+    Issuing findIssuing(ATM atm, double sum) throws ATMExceptions;
+}
